Added lower_s option for signing.
group.cr

Added optional input "lower_s" to signatures, set to true by default. Setting this to false means that "s" will not be modified to be lower or equal to n/2. This can be useful if we want to make sure that the correct public key can easily be retrieved from a signature (otherwise there can be two possibilities, unless we actually know if s was modified when computing the signature).

Added possibility to feed a signature string into recover_public_key()

ecdsa.cr

Bug fix in verify_ethereum_signature() - would not handle signature strings without leading "0x" correctly